Move insert inline formula and graphic commands to shared folder.
