+++ src/article/editor/EditorPackage.js
@@ -54,9 +54,9 @@ import InsertFigureTool from './InsertFigureTool'
 import InsertFigurePanelTool from '../shared/InsertFigurePanelTool'
 import InsertFootnoteCommand from '../shared/InsertFootnoteCommand'
 import InsertFootnoteCrossReferenceCommand from '../shared/InsertFootnoteCrossReferenceCommand'
-import InsertInlineFormulaCommand from './InsertInlineFormulaCommand'
-import InsertInlineGraphicCommand from './InsertInlineGraphicCommand'
-import InsertInlineGraphicTool from './InsertInlineGraphicTool'
+import InsertInlineFormulaCommand from '../shared/InsertInlineFormulaCommand'
+import InsertInlineGraphicCommand from '../shared/InsertInlineGraphicCommand'
+import InsertInlineGraphicTool from '../shared/InsertInlineGraphicTool'
 import { CreateListCommand, ChangeListTypeCommand } from './ListCommands'
 import InsertNodeFromWorkflowCommand from './InsertNodeFromWorkflowCommand'
 import {
